Add `row` and `column` tiling_types

+++ phylanx/plugins/dist_matrixops/dist_constant.hpp
@@ -49,7 +49,7 @@ namespace phylanx { namespace dist_matrixops { namespace primitives
         template <typename T>
         execution_tree::primitive_argument_type constant1d_helper(
             execution_tree::primitive_argument_type&& value,
-            std::size_t const& dim, std::uint32_t const& tile_idx,
+            std::uint32_t const& dim, std::uint32_t const& tile_idx,
             std::uint32_t const& numtiles, std::string&& given_name,
             std::string const& tiling_type, std::string const& name,
             std::string const& codename) const;
